move MergingRun to own module
This allows us to reference MergingRun from other modules, in particular
a future MergingTree.

This commit only moves the code and tweaks formatting and comments.
